refactor: complete dashboard architecture and design system

+++ lib/features/statistics/domain/models/statistics_summary.dart
@@ -13,7 +13,7 @@ class StatisticsSummary {
 
   final int totalCompletions;
 
-  final int completionRate;
+  final double completionRate;
 
   final List<DailyStatistics> weekly;
